requirements.txt
numpy
scipy
scikit-image
opencv-python
pathlib
torch
torchvision
trimesh
chumpy
configer
configargparse
torchgeometry
PyYAML
tqdm
pillow